Modularize app and harden layout and metadata
