+++ index.html
@@ -4,10 +4,39 @@
     <meta charset="utf-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1" />
     <title>AniRanker</title>
+    <meta
+      name="description"
+      content="Curate a personal anime library, compare waifus, and track your ratings with AniRanker."
+    />
+    <link rel="canonical" href="https://aniranker.app/" />
+    <meta property="og:type" content="website" />
+    <meta property="og:title" content="AniRanker" />
+    <meta
+      property="og:description"
+      content="Curate a personal anime library, compare waifus, and track your ratings with AniRanker."
+    />
+    <meta property="og:url" content="https://aniranker.app/" />
+    <meta property="og:image" content="https://aniranker.app/social-preview.png" />
+    <meta property="og:image:alt" content="AniRanker hero showing waifu rankings and anime grid." />
+    <meta name="twitter:card" content="summary_large_image" />
+    <meta name="twitter:title" content="AniRanker" />
+    <meta
+      name="twitter:description"
+      content="Curate a personal anime library, compare waifus, and track your ratings with AniRanker."
+    />
+    <meta name="twitter:image" content="https://aniranker.app/social-preview.png" />
     <link rel="stylesheet" href="styles.css" />
-    <script crossorigin src="https://unpkg.com/react@18/umd/react.production.min.js"></script>
-    <script crossorigin src="https://unpkg.com/react-dom@18/umd/react-dom.production.min.js"></script>
-    <script defer src="app.js"></script>
+    <script
+      crossorigin="anonymous"
+      integrity="sha384-MhvFknC0AhqKZXwSCyPeIirpE1IhTafmLXTvTCai5pxjAqFcqEAAqKH/2lXrFVPn"
+      src="https://unpkg.com/react@18/umd/react.production.min.js"
+    ></script>
+    <script
+      crossorigin="anonymous"
+      integrity="sha384-VQePqH/QE94Buf4g/4MI5NrqDYjdqT7JW6KzhieQUlw6w7U7VD59fPqCAunpXEu6"
+      src="https://unpkg.com/react-dom@18/umd/react-dom.production.min.js"
+    ></script>
+    <script type="module" src="src/main.js"></script>
   </head>
   <body>
     <div id="root"></div>
