+++ src/data/waifus.js
@@ -1,0 +1,331 @@
+export const WAIFUS = [
+  {
+    name: "Rias Gremory",
+    animeId: 11617,
+    animeTitle: "High School DxD",
+    traits: ["Royalty", "Demon", "Leader"],
+    hairColor: "Crimson",
+    skinTone: "Fair",
+    eyeColor: "Blue",
+    chest: "Voluptuous",
+    grace: 8,
+    intensity: 9,
+    warmth: 7,
+    mystique: 6,
+    featured: true,
+  },
+  {
+    name: "Akeno Himejima",
+    animeId: 11617,
+    animeTitle: "High School DxD",
+    traits: ["Priestess", "Demon", "Charmer"],
+    hairColor: "Raven",
+    skinTone: "Fair",
+    eyeColor: "Violet",
+    chest: "Voluptuous",
+    grace: 9,
+    intensity: 8,
+    warmth: 6,
+    mystique: 7,
+    featured: false,
+  },
+  {
+    name: "Zero Two",
+    animeId: 99423,
+    animeTitle: "Darling in the Franxx",
+    traits: ["Pilot", "Hybrid", "Rebel"],
+    hairColor: "Pastel Pink",
+    skinTone: "Fair",
+    eyeColor: "Turquoise",
+    chest: "Balanced",
+    grace: 7,
+    intensity: 9,
+    warmth: 6,
+    mystique: 8,
+    featured: true,
+  },
+  {
+    name: "Asuna Yuuki",
+    animeId: 11757,
+    animeTitle: "Sword Art Online",
+    traits: ["Swordswoman", "Leader", "Reliable"],
+    hairColor: "Chestnut",
+    skinTone: "Fair",
+    eyeColor: "Hazel",
+    chest: "Balanced",
+    grace: 9,
+    intensity: 8,
+    warmth: 8,
+    mystique: 6,
+    featured: true,
+  },
+  {
+    name: "Rem",
+    animeId: 21355,
+    animeTitle: "Re:Zero",
+    traits: ["Demon", "Devoted", "Support"],
+    hairColor: "Sky Blue",
+    skinTone: "Fair",
+    eyeColor: "Azure",
+    chest: "Petite",
+    grace: 8,
+    intensity: 7,
+    warmth: 10,
+    mystique: 6,
+    featured: true,
+  },
+  {
+    name: "Mikasa Ackerman",
+    animeId: 16498,
+    animeTitle: "Attack on Titan",
+    traits: ["Soldier", "Protector", "Stoic"],
+    hairColor: "Jet Black",
+    skinTone: "Olive",
+    eyeColor: "Gray",
+    chest: "Athletic",
+    grace: 8,
+    intensity: 10,
+    warmth: 6,
+    mystique: 7,
+    featured: true,
+  },
+  {
+    name: "Yor Forger",
+    animeId: 140960,
+    animeTitle: "SPY x FAMILY",
+    traits: ["Assassin", "Family", "Graceful"],
+    hairColor: "Raven",
+    skinTone: "Porcelain",
+    eyeColor: "Red",
+    chest: "Balanced",
+    grace: 9,
+    intensity: 8,
+    warmth: 7,
+    mystique: 8,
+    featured: true,
+  },
+  {
+    name: "Nezuko Kamado",
+    animeId: 101922,
+    animeTitle: "Demon Slayer",
+    traits: ["Demon", "Protector", "Pure"],
+    hairColor: "Black Ombre",
+    skinTone: "Porcelain",
+    eyeColor: "Rose",
+    chest: "Petite",
+    grace: 8,
+    intensity: 7,
+    warmth: 9,
+    mystique: 7,
+    featured: false,
+  },
+  {
+    name: "Marin Kitagawa",
+    animeId: 132405,
+    animeTitle: "My Dress-Up Darling",
+    traits: ["Cosplayer", "Sunshine", "Creative"],
+    hairColor: "Honey Blonde",
+    skinTone: "Fair",
+    eyeColor: "Gold",
+    chest: "Balanced",
+    grace: 8,
+    intensity: 6,
+    warmth: 9,
+    mystique: 5,
+    featured: false,
+  },
+  {
+    name: "Makima",
+    animeId: 127230,
+    animeTitle: "Chainsaw Man",
+    traits: ["Leader", "Manipulator", "Calm"],
+    hairColor: "Copper",
+    skinTone: "Fair",
+    eyeColor: "Amber",
+    chest: "Balanced",
+    grace: 9,
+    intensity: 9,
+    warmth: 4,
+    mystique: 10,
+    featured: false,
+  },
+  {
+    name: "Kaguya Shinomiya",
+    animeId: 101921,
+    animeTitle: "Love is War",
+    traits: ["Heiress", "Strategist", "Elegant"],
+    hairColor: "Ink Black",
+    skinTone: "Porcelain",
+    eyeColor: "Ruby",
+    chest: "Balanced",
+    grace: 9,
+    intensity: 7,
+    warmth: 6,
+    mystique: 8,
+    featured: false,
+  },
+  {
+    name: "Nobara Kugisaki",
+    animeId: 113415,
+    animeTitle: "Jujutsu Kaisen",
+    traits: ["Sorcerer", "Fierce", "Stylish"],
+    hairColor: "Copper Brown",
+    skinTone: "Olive",
+    eyeColor: "Amber",
+    chest: "Athletic",
+    grace: 8,
+    intensity: 9,
+    warmth: 6,
+    mystique: 7,
+    featured: false,
+  },
+  {
+    name: "Albedo",
+    animeId: 20832,
+    animeTitle: "Overlord",
+    traits: ["Guardian", "Demon", "Devoted"],
+    hairColor: "Raven",
+    skinTone: "Porcelain",
+    eyeColor: "Amber",
+    chest: "Voluptuous",
+    grace: 9,
+    intensity: 8,
+    warmth: 7,
+    mystique: 9,
+    featured: false,
+  },
+  {
+    name: "Rei Ayanami",
+    animeId: 30,
+    animeTitle: "Neon Genesis Evangelion",
+    traits: ["Pilot", "Enigmatic", "Calm"],
+    hairColor: "Icy Blue",
+    skinTone: "Fair",
+    eyeColor: "Red",
+    chest: "Petite",
+    grace: 8,
+    intensity: 6,
+    warmth: 4,
+    mystique: 10,
+    featured: false,
+  },
+  {
+    name: "Hitagi Senjougahara",
+    animeId: 17074,
+    animeTitle: "Monogatari Series",
+    traits: ["Tsundere", "Sharp", "Poised"],
+    hairColor: "Indigo",
+    skinTone: "Porcelain",
+    eyeColor: "Magenta",
+    chest: "Balanced",
+    grace: 9,
+    intensity: 7,
+    warmth: 5,
+    mystique: 8,
+    featured: false,
+  },
+  {
+    name: "Kurisu Makise",
+    animeId: 9253,
+    animeTitle: "Steins;Gate",
+    traits: ["Scientist", "Sharp", "Compassionate"],
+    hairColor: "Auburn",
+    skinTone: "Fair",
+    eyeColor: "Hazel",
+    chest: "Balanced",
+    grace: 8,
+    intensity: 7,
+    warmth: 8,
+    mystique: 7,
+    featured: false,
+  },
+  {
+    name: "Rukia Kuchiki",
+    animeId: 269,
+    animeTitle: "Bleach",
+    traits: ["Soul Reaper", "Calm", "Artist"],
+    hairColor: "Raven",
+    skinTone: "Fair",
+    eyeColor: "Violet",
+    chest: "Petite",
+    grace: 8,
+    intensity: 8,
+    warmth: 6,
+    mystique: 7,
+    featured: false,
+  },
+  {
+    name: "Nami",
+    animeId: 21,
+    animeTitle: "One Piece",
+    traits: ["Navigator", "Strategist", "Bold"],
+    hairColor: "Tangerine",
+    skinTone: "Sun-kissed",
+    eyeColor: "Brown",
+    chest: "Voluptuous",
+    grace: 7,
+    intensity: 7,
+    warmth: 8,
+    mystique: 6,
+    featured: false,
+  },
+  {
+    name: "Hinata Hyuga",
+    animeId: 20,
+    animeTitle: "Naruto",
+    traits: ["Ninja", "Gentle", "Resilient"],
+    hairColor: "Midnight",
+    skinTone: "Fair",
+    eyeColor: "Lavender",
+    chest: "Balanced",
+    grace: 9,
+    intensity: 6,
+    warmth: 9,
+    mystique: 6,
+    featured: false,
+  },
+  {
+    name: "Shiki Ryougi",
+    animeId: 6954,
+    animeTitle: "The Garden of Sinners",
+    traits: ["Mystic", "Assassin", "Stoic"],
+    hairColor: "Midnight",
+    skinTone: "Porcelain",
+    eyeColor: "Blue",
+    chest: "Athletic",
+    grace: 9,
+    intensity: 9,
+    warmth: 4,
+    mystique: 10,
+    featured: false,
+  },
+];
+
+export const WAIFU_ATTRIBUTE_LABELS = {
+  name: "Name",
+  grace: "Grace",
+  intensity: "Intensity",
+  warmth: "Warmth",
+  mystique: "Mystique",
+};
+
+export function sortWaifuList(list, key, direction) {
+  const sorted = list.slice();
+  if (key === "name") {
+    sorted.sort((a, b) =>
+      direction === "asc" ? a.name.localeCompare(b.name) : b.name.localeCompare(a.name)
+    );
+    return sorted;
+  }
+
+  sorted.sort((a, b) => {
+    const left = typeof a[key] === "number" ? a[key] : 0;
+    const right = typeof b[key] === "number" ? b[key] : 0;
+    if (right === left) {
+      return a.name.localeCompare(b.name);
+    }
+    return direction === "asc" ? left - right : right - left;
+  });
+
+  return sorted;
+}
