+++ src/main.js
@@ -1,0 +1,7 @@
+import App from './components/App.js';
+
+const rootElement = document.getElementById('root');
+if (rootElement) {
+  const root = ReactDOM.createRoot(rootElement);
+  root.render(React.createElement(App));
+}
